Reorganize directory structure
HandlerMiddleware really belongs in the Handler folder and Middleware
is now a core concept so it should sit with the other master
interfaces.

The plugins dir is a bit of lame naming but it expresses the intent and
we can expand it later on with other one-off default plugins.
